pyproject.toml
[project]
name = "test-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.19",
    "docling>=2.48.0",
    "dotenv>=0.9.9",
    "git-filter-repo>=2.47.0",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.1",
    "llama-cpp-python>=0.3.16",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "simplegmail>=4.1.1",
]
